requirements.txt
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
idna==3.4
python-telegram-bot==20.3
requests==2.30.0
sniffio==1.3.0
urllib3==2.0.2
